重构 runtime 管理及状态同步逻辑：
- 引入 `RuntimeManager`，重构 runtime 状态管理和任务控制流程
- 替换 `LocalRuntimeRegistry` 为新的 `RuntimeReadModel` 接口，清理遗留逻辑
- 优化 session 处理，统一引入 `generation` 校验以确保状态一致性
- 调整事件流转，引入 internal event 模型，消除长 await 操作

+++ crates/media-agent/src/heartbeat.rs
@@ -169,8 +169,8 @@ fn sample_disk(path: &str) -> Option<DiskSample> {
     }
 
     let stat = unsafe { stat.assume_init() };
-    let total = stat.f_blocks.saturating_mul(stat.f_frsize);
-    let free = stat.f_bavail.saturating_mul(stat.f_frsize);
+    let total = (stat.f_blocks as u64).saturating_mul(stat.f_frsize as u64);
+    let free = (stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64);
     if total == 0 {
         return Some(DiskSample::default());
     }
